requirements.txt
colorama==0.4.6
dnspython==2.7.0
tqdm==4.67.1
aiohttp==3.11.9
pyvis==0.3.2
scrapy==2.12.0
zaproxy==0.3.2
python-dotenv==1.0.1
langchain==0.2.16
langchain-groq==0.1.10
fpdf2==2.8.1
langchain-community==0.2.13
langchain-huggingface==0.0.3
faiss-cpu==1.9.0